requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
biopython=1.73=py37h1de35cc_0
blas=1.0=mkl
bzip2=1.0.6=h1de35cc_1002
ca-certificates=2018.11.29=ha4d7672_0
certifi=2018.11.29=py37_1000
intel-openmp=2019.1=144
libcxx=7.0.0=h2d50403_2
libffi=3.2.1=h0a44026_1005
libgfortran=3.0.1=h93005f0_2
llvm-meta=7.0.0=0
mkl=2019.1=144
mkl_fft=1.0.10=py37h1de35cc_1
mkl_random=1.0.2=py37h1702cab_2
ncurses=6.1=h0a44026_1002
numpy=1.15.4=py37hacdab7b_0
numpy-base=1.15.4=py37h6575580_0
openssl=1.0.2p=h1de35cc_1002
pip=19.0.1=py37_0
python=3.7.1=h145921a_1000
readline=7.0=hcfe32e1_1001
setuptools=40.7.1=py37_0
sqlite=3.26.0=h1765d9f_1000
tk=8.6.9=ha441bb4_1000
wheel=0.32.3=py37_0
xz=5.2.4=h1de35cc_1001
zlib=1.2.11=h1de35cc_1004
